Create basic linked_list
Added two classes for a linked list: first the Node class, that
represents a part of a linked list, and Linked_list, the class that
can be empty or store many nodes.

There are also methods to access and modify the value of a node, the
next_node and the head node of a Linked_list.