Add basic method to LinkedList class
Some of the basic operations for this class are to get the tail of the
linked list (last node of the list) and append a node at the end of
the list. Also another one is to get the head of the linked list (the
first node of the list), but this method was added when the LinkedList
class was created.