Add prepend method to LinkedList
The prepend operation adds a new node at the start of the list. If this
method is called on an empty list, then call the append method, if not
just create a temporary node to store the previous head node, then
create and assing the new node to the head, and finally add the previous
head node to the next_node of the current head node.