Add contains? method to LinkedList class
The contains? method checks if the value passed is on the linked list.
If it is then it returns true, if not it returns false. It first checks
the head node and its next node values and then it iterates through all
the remaining nodes until it founds the value or there are no more
nodes.