Add insert_at method
This method receives a value to be inserted in the linked list on a
given index. If the index is greater but not the next index it returns
nil. If the index is 0, then it calls the prepend method to assing it
at the head. If the index is equal to the size of the linked list, then
it calls append to assign it at the tail of the list.

When the index given is nor head nor tail, then it creates the new node,
then it finds the previous node given the index and the actual node
with the given index; after this, it assigns to the next_node of the
previous node the new node and for the next_node of this new node it
sets the other node to avoid deleting it from the list.